Implementing redux on admin pages
