+++ webapp/src/app/view/admin/beagle-box/admin-beagle-box.jsx
@@ -1,0 +1,18 @@
+import React, { Component } from 'react'
+import { Row, Col } from 'react-flexbox-grid'
+
+
+export default class AdminBeagleBox extends Component {
+  render() {
+    return (
+      <Row className="row-fluid">
+        <Col className="col-fluid" md={12} sm={12} xs={12}>
+
+        </Col>
+        <Col className="col-fluid" md={12} sm={12} xs={12}>
+
+        </Col>
+      </Row>
+    );
+  }
+}
