+++ webapp/src/app/view/admin/welcome/index.js
@@ -1,20 +1,18 @@
-import React, { Component } from 'react'
-import { Row, Col } from 'react-flexbox-grid'
+import { connect } from "react-redux"
+import AdminWelcomeComponent from "./admin-welcome"
 
-import assets from '../assets'
+const stateToProps = (state) => {
+  return {
 
-export default class AdminWelcome extends Component {
-  render() {
-    return (
-      <Row className="row-fluid">
-        <Col className="col-fluid col-image" md={12} sm={12} xs={12}>
-          <img className="img-admin" src={assets.admin} alt="Ícone de Adminstrador" />
-        </Col>
-        <Col className="col-fluid" md={12} sm={12} xs={12}>
-          <h1>Bem-vindo(a)</h1>
-          <h2>Área exclusiva para administração</h2>
-        </Col>
-      </Row>
-    );
   }
-}
+};
+
+const dispatchToProps = (dispatch) => {
+  return {
+
+  }
+};
+
+const AdminWelcome = connect(stateToProps, dispatchToProps)(AdminWelcomeComponent);
+
+export default AdminWelcome;
